Implement Attributable trait
