+++ src/Illuminate/Auth/GenericUser.php
@@ -1,15 +1,11 @@
 <?php namespace Illuminate\Auth;
 
 use Illuminate\Contracts\Auth\Authenticatable as UserContract;
+use Illuminate\Support\Traits\Attributable;
 
 class GenericUser implements UserContract {
 
-	/**
-	 * All of the user's attributes.
-	 *
-	 * @var array
-	 */
-	protected $attributes;
+	use Attributable;
 
 	/**
 	 * Create a new generic User object.
@@ -71,51 +67,6 @@ class GenericUser implements UserContract {
 	public function getRememberTokenName()
 	{
 		return 'remember_token';
-	}
-
-	/**
-	 * Dynamically access the user's attributes.
-	 *
-	 * @param  string  $key
-	 * @return mixed
-	 */
-	public function __get($key)
-	{
-		return $this->attributes[$key];
-	}
-
-	/**
-	 * Dynamically set an attribute on the user.
-	 *
-	 * @param  string  $key
-	 * @param  mixed   $value
-	 * @return void
-	 */
-	public function __set($key, $value)
-	{
-		$this->attributes[$key] = $value;
-	}
-
-	/**
-	 * Dynamically check if a value is set on the user.
-	 *
-	 * @param  string  $key
-	 * @return bool
-	 */
-	public function __isset($key)
-	{
-		return isset($this->attributes[$key]);
-	}
-
-	/**
-	 * Dynamically unset a value on the user.
-	 *
-	 * @param  string  $key
-	 * @return void
-	 */
-	public function __unset($key)
-	{
-		unset($this->attributes[$key]);
 	}
 
 }
